Implement creation of configuration records from the web interface. Fixes #2 Implement removal of configuration records from the web interface. Fixes #3 Remove hard coded config_id. Fixes #4 Fix indenting.

+++ web-configuration/users/delete.php
@@ -4,15 +4,18 @@
 include("../conf/config.php");
 
 //getting user_id of the data from url
-$user_id = mysqli_real_escape_string($mysqli, $_GET['user_id']);
-
-//deleting the row from table
-$result = mysqli_query($mysqli, "DELETE FROM users WHERE user_id='$user_id'");
-mysqli_close($mysqli);
-if($result)
+if (isset($_GET['user_id']))
 {
-				//redirecting to the display page (index.php in our case)
-				header("Location:index.php");
+  $user_id = mysqli_real_escape_string($mysqli, $_GET['user_id']);
+
+  //deleting the row from table
+  $result = mysqli_query($mysqli, "DELETE FROM users WHERE user_id='$user_id'");
+  mysqli_close($mysqli);
+  if($result)
+  {
+          //redirecting to the display page (index.php in our case)
+          header("Location:index.php");
+  }
+  else {print_r(mysqli_error_list($mysqli));}
 }
-else {print_r(mysqli_error_list($mysqli));}
 ?>
